feat: include fee transfer logs for reverting transactions (#729)
This PR ensures that logs for fee token transfer are always included
into the transaction receipt, even if transaction reverted.

This work arounds the fact that revm `ExecutionResult` does not contain
`logs` for `Revert`/`Halt` variants. We should remove this once it's
supported

This also contains a fix for a bug when no `Transfer` log is emitted for
cases when there's no fee token refund happening