fix: set coinbase validator token to LinkingUSD (#910)
Note: most of the diff is `test-genesis-moderato.json` formatting,
toggle "Hide whitespace` to make it load

Changes in the PR:
- `generate-genesis` command is changed to set the `coinbase`'s fee
token to `LinkingUSD`
- before it was not set for the `coinbase` at all, and was instead
configured just for `Address::ZERO` and vaidators admin EOAs that were
not used as `feeRecipient`s
- node launching in tests is changed to preserve `coinbase` value from
genesis for new blocks. before it was set to random address which didn't
preserve the validator fee token config
- `test-genesis`/`test-genesis-moderato` are changed to use a new random
address as coinbase instead of the first sender of test mnemonic. this
is needed because this address is now actually used as validator for all
other blocks, and was messing up some tests that were relying on user
and validator addresses being different
- tests are adjusted
- most of the adjustments are just changing `DEFAULT_FEE_TOKEN` to
`LINKING_USD_ADDRESS`
- some of the tests no longer work pre-Moderato because they require
proper FeeAMM accounting which was fixed in
https://github.com/tempoxyz/tempo/pull/892