docs(specs): change DEX pair storage layout to be order-sensitive (#1791)
This changes how DEX pairs are keyed so that a pair that uses token A as
a base token and token B as a quote token would be keyed differently
from a pair that uses token B as a base token and token A as a quote
token. This would be a breaking change that affects storage layout and
cannot be implemented in a hard fork without a regenesis.

---------

Co-authored-by: 0xKitsune <77890308+0xKitsune@users.noreply.github.com>