fix(specs): check 403 policy on both tokens in stablecoin DEX (#1805)
Partially addresses TMPO-20.

Currently, the DEX only checks TIP-403 policies on the token that is
being escrowed (i.e. sold by the limit order). This could allow a maker
to place an order to buy a token that they are not authorized to
transfer. This could have slightly odd results in the case of
blacklisting, but particularly bad UX in the case of whitelisting, where
anyone would be allowed to buy a token without being on the whitelist
for it, resulting in them being stuck with the token.

This PR changes this in the reference implementation by having the
placeOrder function check the authorization policies on both tokens. It
also checks that the DEX itself is authorized by both tokens.

In the case of a flip order, this will cause the newly placed order to
silently fail (ending the cycle of flips).

---------

Co-authored-by: 0xKitsune <77890308+0xKitsune@users.noreply.github.com>
Co-authored-by: 0xKitsune <0xkitsune@protonmail.com>